fix mistral ocr
