+++ engine/llm_services/utils.py
@@ -92,3 +92,43 @@ def check_usage(provider: str) -> None:
                 "Please provide your own key."
             )
     return None
+
+
+def make_mistral_ocr_compatible(payload_json: dict) -> list[dict]:
+
+    if "messages" not in payload_json or not payload_json["messages"]:
+        return None
+
+    for message in reversed(payload_json["messages"]):
+        if not isinstance(message, dict) or "content" not in message:
+            continue
+
+        content = message["content"]
+        if not content:
+            continue
+
+        payload_json["messages"][-1]["content"] = content if isinstance(content, list) else [content]
+
+        for content_item in payload_json["messages"][-1]["content"]:
+            if (
+                isinstance(content_item, dict)
+                and "file" in content_item
+                and isinstance(content_item["file"], dict)
+                and "file_data" in content_item["file"]
+            ):
+                return {
+                    "type": "document_url",
+                    "document_url": f"data:application/pdf;base64,{content_item['file']['file_data']}",
+                }
+            elif (
+                isinstance(content_item, dict)
+                and "image_url" in content_item
+                and isinstance(content_item["image_url"], dict)
+                and "url" in content_item["image_url"]
+            ):
+                return {
+                    "type": "image_url",
+                    "image_url": content_item["image_url"]["url"],
+                }
+
+    return None
